Refactoring: split tests to suits
